// src/types/ui/enums/Type.ts
export enum Type {
    SELECTION_WHEEL = "selection_wheel",
    PANEL = "panel",
    SCREEN = "screen",
    STACK_PANEL = "stack_panel",
    LABEL = "label",
    IMAGE = "image",
    INPUT_PANEL = "input_panel",
    CUSTOM = "custom",
    GRID = "grid",
    FACTORY = "factory",
    BUTTON = "button",
    TOGGLE = "toggle",
    SLIDER = "slider",
    EDIT_BOX = "edit_box",
    DROPDOWN = "dropdown",
    SCROLL_VIEW = "scroll_view",
    SLIDER_BOX = "slider_box",
    SCROLLBAR_BOX = "scrollbar_box",
    SCROLL_TRACK = "scroll_track",
    GRID_PAGE_INDICATOR = "grid_page_indicator",
    IMAGE_CYCLER = "image_cycler",
    LABEL_CYCLER = "label_cycler",
    COLLECTION_PANEL = "collection_panel",
    TOOLTIP_TRIGGER = "tooltip_trigger",
    TAB = "tab",
    CAROUSEL_LABEL = "carousel_label",
    COMBOX = "combox",
    LAYOUT = "layout",
    STACK_GRID = "stack_grid",
    JOYSTICK = "joystick",
    RICH_TEXT = "rich_text",
    SIXTEEN_NINE_LAYOUT = "sixteen_nine_layout",
    MUL_LINES = "mul_lines",
    UNKNOWN = "unknown",
}